# Reduce the size of the Cookie by 32 bytes.